+++ grafana-plugin/e2e-tests/integrations/heartbeat.test.ts
@@ -1,32 +1,25 @@
-import { test, Page, expect, Locator } from '../fixtures';
+import { test, Page, expect } from '../fixtures';
 
 import { generateRandomValue, selectDropdownValue } from '../utils/forms';
 import { createIntegration } from '../utils/integrations';
 
+const HEARTBEAT_SETTINGS_FORM_TEST_ID = 'heartbeat-settings-form';
+
 test.describe("updating an integration's heartbeat interval works", async () => {
-  test.slow();
-
-  const _openIntegrationSettingsPopup = async (page: Page): Promise<Locator> => {
-    const integrationSettingsPopupElement = page.getByTestId('integration-settings-context-menu');
-    await integrationSettingsPopupElement.click();
-    return integrationSettingsPopupElement;
-  };
-
   const _openHeartbeatSettingsForm = async (page: Page) => {
-    const integrationSettingsPopupElement = await _openIntegrationSettingsPopup(page);
-
+    const integrationSettingsPopupElement = page.getByTestId('integration-settings-context-menu');
+    await integrationSettingsPopupElement.waitFor({ state: 'visible' });
     await integrationSettingsPopupElement.click();
 
     await page.getByTestId('integration-heartbeat-settings').click();
   };
 
   test('change heartbeat interval', async ({ adminRolePage: { page } }) => {
-    const integrationName = generateRandomValue();
-    await createIntegration(page, integrationName);
+    await createIntegration(page, generateRandomValue());
 
     await _openHeartbeatSettingsForm(page);
 
-    const heartbeatSettingsForm = page.getByTestId('heartbeat-settings-form');
+    const heartbeatSettingsForm = page.getByTestId(HEARTBEAT_SETTINGS_FORM_TEST_ID);
 
     const value = '30 minutes';
 
@@ -49,23 +42,24 @@ test.describe("updating an integration's heartbeat interval works", async () => 
     expect(heartbeatIntervalValue).toEqual(value);
   });
 
-  test('send heartbeat', async ({ request, adminRolePage: { page } }) => {
-    const integrationName = generateRandomValue();
-    await createIntegration(page, integrationName);
+  test('send heartbeat', async ({ adminRolePage: { page } }) => {
+    await createIntegration(page, generateRandomValue());
 
     await _openHeartbeatSettingsForm(page);
 
-    const heartbeatSettingsForm = page.getByTestId('heartbeat-settings-form');
+    const heartbeatSettingsForm = page.getByTestId(HEARTBEAT_SETTINGS_FORM_TEST_ID);
 
     const endpoint = await heartbeatSettingsForm
       .getByTestId('input-wrapper')
       .locator('input[class*="input-input"]')
       .inputValue();
 
-    await request.get(endpoint);
+    /**
+     * make an HTTP call to the integration's hearbeat URL and assert that it was called
+     * (ie. the greenheart badge is shown)
+     */
+    await page.request.get(endpoint);
     await page.reload({ waitUntil: 'networkidle' });
-
-    // If heartbeat was never sent, there will be no badge
     await page.getByTestId('heartbeat-badge').waitFor({ state: 'visible' });
   });
 });